requirements.txt
astroid==2.4.1
bleach==3.1.5
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
defusedxml==0.6.0
diff-match-patch==20181111
dj-static==0.0.6
Django==2.2.7
django-admin-honeypot==1.1.0
django-allauth==0.41.0
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-countries==6.1.2
django-crispy-forms==1.8.1
django-debug-toolbar==2.2
django-embed-video==1.3.2
django-import-export==2.1.0
django-js-asset==1.2.2
django-markdownify==0.8.0
django-memcache-status==2.2
django-rest-auth==0.9.5
django-taggit==1.2.0
django-typeahead==1.0.1
djangorestframework==3.11.0
et-xmlfile==1.0.1
get==2019.4.13
gunicorn==20.0.4
idna==2.8
isort==4.3.21
jdcal==1.4.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupPy==1.14
matplotlib==3.1.1
mccabe==0.6.1
numpy==1.17.3
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.1
packaging==20.3
Pillow==7.0.0
pipenv==2018.11.26
post==2019.4.13
prettytable==0.7.2
psycopg2==2.8.4
public==2019.4.13
pygal==2.4.0
pygal-maps-world==1.0.2
pygame==1.9.6
pylint==2.5.2
pyparsing==2.4.3
PyPDF2==1.26.0
pyperclip==1.7.0
python-dateutil==2.8.1
python-gettext==4.0
python-memcached==1.59
python3-memcached==1.51
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
query-string==2019.4.13
regex==2019.12.20
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.3.0
six==1.12.0
sorl-thumbnail==12.6.3
sqlparse==0.3.0
static3==0.7.0
stripe==2.42.0
tablib==2.0.0
tabulate==0.8.6
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.6
virtualenv==16.7.7
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.5
webencodings==0.5.1
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
